+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - `CHANGELOG`: Add a structure to CHANGELOG logs
+- `tman`: Add support to not switch to newly created project
 
 ### Changed